[#1380] neofs-cli: move `container` command to a separate package
